first working version of datahandler obejct
it has the power of taking a pandas dataframe with a number of
categorical features and onehot encode them, returning nicely
formated numpy array for x, y and z varibales.

this is the object necessary to glue the data that can easily be pulled
from the database to the low-level representation necessary to plug
values into the shifted-beta model.